+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Added option to display the DKIM header if when an e-mail with a DKIM signature, SPF or DMARC result is viewed (#462).
 - Authentication-Results header: Invalid headers by Outlook are now accepted if reading of non RFC compliant ARHs is enabled (#423).
 - The DKIM selector is now displayed in the DKIM button pop-up (#510).
+- The options page of the add-on can now be open from the DKIM button pop-up.
 
 ## 6.0.1 (2025-03-10)